Make act.py return floats

+++ act.py
@@ -2,26 +2,26 @@ from math import exp
 
 # SIGMOID
 def sigmoid(x):
-	return 1 / (1 + exp(-x))
+	return float(1 / (1 + exp(-x)))
 
 def dsigmoid(x):
-	return x * (1-x)
+	return float(x * (1-x))
 
 # TANH
 def tanh(x):
 	try:
 		a = (exp(x) - exp(-x))
 		b = (exp(x) + exp(-x))
-		return a / b
+		return float(a / b)
 	except OverflowError:
 		print(x)
 
 def dtanh(x):
-	return 1 - pow(tanh(x), 2)
+	return float(1 - pow(tanh(x), 2))
 
 def relu(x):
 	if (x > 0):
-		return x
+		return float(x)
 	else:
 		return 0.0
